requirements.txt
python-docx
scispacy
spacy
fuzzywuzzy
PyQt5